services/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlmodel==0.0.16
python-dotenv==1.0.1
mysqlclient==2.2.4